Factor checks out of ACAReviewTable class
