+++ sparkles/roll_optimize.py
@@ -339,6 +339,7 @@ class RollOptimizeMixin:
 
         :return: None
         """
+        from sparkles.core import check_catalog
 
         if self.loud:
             print(f" Exploring roll options {method=}")
@@ -359,7 +360,7 @@ class RollOptimizeMixin:
 
         # Special case, first roll option is self but with obsid set to roll
         acar = deepcopy(self)
-        acar.check_catalog()
+        check_catalog(acar)
         acar.is_roll_option = True
         roll_options = [
             {
@@ -414,7 +415,7 @@ class RollOptimizeMixin:
                 acar = self.__class__(aca_rolled, obsid=self.obsid, is_roll_option=True)
 
                 # Do the review and set up messages attribute
-                acar.check_catalog()
+                check_catalog(acar)
 
                 roll_option = {
                     "acar": acar,
